Move types to own folder
Other than moving the types, I made a few other code-style changes as well.
- Moved generated types to `schema.ts`.
- Flagged all `**/generated/` directories (not just the root one).
- Removed default export from `globals.d.ts`.
- Cleaned up `secrets.ts`.
