+++ src/types/context.d.ts
@@ -1,5 +1,5 @@
 import type {JwtPayload} from 'jsonwebtoken'
-import type {UserDbObject} from './generated/types.js'
+import type {UserDbObject} from './generated/schema.js'
 
 export interface Context extends Record<string, unknown>, JwtPayload {
 	user?: UserDbObject,
